v1: `PreciseFlex`: split the package into config / driver / kinematics / arm_backend modules
Restructure only, no behaviour change. The single 2400-line `precise_flex.py` is split into the
per-concern module layout already used by the KX2 backend (`pylabrobot/paa/kx2`):

- `config.py` - the `Axis` enum and the resolved `PreciseFlexConfiguration` record
- `driver.py` - pure transport: the socket I/O connection and device lifecycle
- `kinematics.py` - FK/IK plus the value types it produces/consumes (`PreciseFlexCartesianPose`,
  `WorkingVolume`, the orientation/wrist literals)
- `arm_backend.py` - the capability backend and its controller-reply parsers
- `precise_flex.py` - the thin per-model device front-ends

`tests/` becomes a package, split per module (`kinematics_tests.py`, `arm_backend_tests.py`).

`driver` depends on nothing PreciseFlex-specific; `config` depends only on `kinematics`; there are no
import cycles. Public imports via the package `__init__` are unchanged.
