+++ pylabrobot/brooks/precise_flex/tests/arm_backend_tests.py
@@ -2,7 +2,7 @@ import unittest
 from typing import Tuple
 from unittest.mock import AsyncMock, MagicMock
 
-from pylabrobot.brooks.precise_flex import Axis, PreciseFlex400Backend, kinematics
+from pylabrobot.brooks.precise_flex import Axis, PreciseFlex400Backend
 
 
 def _make_backend(
@@ -18,16 +18,6 @@ def _make_backend(
     closed_gripper_position=closed_gripper_position,
   )
   return backend, driver
-
-
-class TestClassifyPF400Reach(unittest.TestCase):
-  """Link lengths are classified as standard, extended, or unknown reach."""
-
-  def test_classify_pf400_reach(self):
-    self.assertEqual(kinematics._classify_pf400_reach((225, 210)), "standard")
-    self.assertEqual(kinematics._classify_pf400_reach((302, 289)), "extended")
-    self.assertEqual(kinematics._classify_pf400_reach((303, 288)), "extended")  # within tolerance
-    self.assertEqual(kinematics._classify_pf400_reach((500, 500)), "unknown")
 
 
 class TestPreciseFlex400Gripper(unittest.IsolatedAsyncioTestCase):
